+++ app/react/common/stacks/queries/query-keys.ts
@@ -1,6 +1,6 @@
 import { StackId } from '../types';
 
 export const stacksQueryKeys = {
-  stackFile: (stackId: StackId) => ['stacks', stackId, 'file'],
-  stacks: ['stacks'],
+  stackFile: (stackId: StackId) => ['stacks', stackId, 'file'] as const,
+  stacks: ['stacks'] as const,
 };
