+++ app/react/kubernetes/applications/components/RedeployGitAppForm/types.ts
@@ -1,6 +1,5 @@
 import {
   AutoUpdateModel,
-  AutoUpdateResponse,
   GitAuthModel,
   GitCredentialsModel,
 } from '@/react/portainer/gitops/types';
@@ -11,12 +10,6 @@ export interface KubeAppGitFormValues {
   repositoryURL: string;
   tlsSkipVerify: boolean;
   autoUpdate: AutoUpdateModel;
-}
-
-export interface UpdateKubeGitStackPayload extends GitCredentialsModel {
-  AutoUpdate: AutoUpdateResponse | null;
-  RepositoryReferenceName: string;
-  TLSSkipVerify: boolean;
 }
 
 export interface RedeployGitStackPayload extends GitCredentialsModel {
